works as before but in fragments
